+++ README.md
@@ -2,7 +2,7 @@
 
 
 <div align="center">
-  <img src="src/Cizzla_logo.png" alt="logo" width="140"  height="auto" />
+  <img src="src/images/Cizzla_logo.png" alt="logo" width="140"  height="auto" />
   <br/>
 
   <h3><b>To-do List README Template</b></h3>
