+++ src/images/reset.png
