+++ src/index.html
@@ -6,6 +6,7 @@
   <title>Todo List</title>
   <link rel="stylesheet" href="style.css">
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/css/all.min.css">
+  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.10.5/font/bootstrap-icons.css">
 </head>
 <body>
     <div class="todo-container">
@@ -13,7 +14,7 @@
         <div class="heading">
             <div class="heading-content">
                 <h3>Today's To Do</h1>
-                <img class="resetImg" src="reset.png" width="20px">
+                <i class="bi bi-arrow-repeat"></i>
             </div>
              <hr>
         </div><hr>
@@ -21,7 +22,7 @@
         <form id="todo-form">
             <div class="form-container">
                 <input type="text" id="task-input" placeholder="Add to your list..." required>
-                <img class="enterImg" src="enter-3552033.svg" width="20px">
+                <i class="bi bi-arrow-return-left"></i>
             </div>
             <hr>
         </form>
