+++ src/style.css
@@ -22,6 +22,11 @@ h3 {
   color: rgb(70, 67, 67);
 }
 
+.fa-arrows-rotate {
+  color: red;
+  width: 20px;
+}
+
 hr {
   color: red;
 }
@@ -32,7 +37,7 @@ hr {
 }
 
 .heading-content {
-  width: 98%;
+  width: 99%;
   display: flex;
   align-items: center;
   justify-content: space-between;
