+++ webpack.config.js
@@ -1,16 +1,32 @@
+const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
   entry: './src/index.js',
   output: {
-    path: `${__dirname}/dist`,
     filename: 'bundle.js',
+    path: path.resolve(__dirname, 'dist'),
+    clean: true,
   },
   module: {
     rules: [
       {
+        test: /\.js$/,
+        exclude: /node_modules/,
+        use: {
+          loader: 'babel-loader',
+          options: {
+            presets: ['@babel/preset-env'],
+          },
+        },
+      },
+      {
         test: /\.css$/,
         use: ['style-loader', 'css-loader'],
+      },
+      {
+        test: /\.(png|svg)$/i,
+        type: 'asset/resource',
       },
     ],
   },
@@ -20,6 +36,9 @@ module.exports = {
     }),
   ],
   devServer: {
-    contentBase: './dist',
+    static: path.resolve(__dirname, 'dist'),
+    compress: true,
+    port: 8080,
+    open: true,
   },
 };
